Factor out Waypoint GPX conversions
